put chart related utilities to pipelines/generation/utils/chart.py
